backend/mainApp/serializers.py
from rest_framework import serializers
from .models import CustomUser, Booking, Room, LeisureItem, Transaction, TransactionItem

class CustomUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = '__all__'
        extra_kwargs = {
            'password': {'write_only': True},
            'lastAccessed': {'read_only': True},
        }

class BookingSerializer(serializers.ModelSerializer):
    class Meta:
        model = Booking
        fields = '__all__'        

class RoomSerializer(serializers.ModelSerializer):
    class Meta:
        model = Room
        fields = '__all__'

class LeisureItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = LeisureItem
        fields = '__all__'     

class TransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transaction
        fields = '__all__'     

class TransactionItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = TransactionItem
        fields = '__all__'  