requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
